+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed startup order of `Tornex.Scheduler.QueryRegistry` in `Tornex.Scheduler.Supervisor`
 
 ## [0.6.0] - 2026-06-05
 ### Added